move profiler in its own folder
